requirements.txt
streamlit 
langchain
langchain-google-genai
google-generativeai
pandas
python-dotenv